Simple API with ninja (#1)
* init ninja

* startapp expense_tracker

* model

* feat: Add 'expense_tracker' app to INSTALLED_APPS

The 'expense_tracker' app has been added to the list of installed apps in the Django project settings. This allows the app to be recognized and used within the project.

* chore: Add initial migration for 'expense_tracker' app

The initial migration file for the 'expense_tracker' app has been added. This migration includes the creation of the 'User' and 'Expense' models with their respective fields.

* chore: Update .gitignore to exclude .vscode directory

* chore: Register 'Expense' model in admin site

The 'Expense' model has been registered in the admin site of the 'expense_tracker' app. This allows the model to be managed and viewed through the Django admin interface.

* update model to use system's user

* migration

* fix model

* model: change date to datetime

* migration

* chore: Update expense_tracker API to create and list expenses

The expense_tracker API has been updated to include the functionality to create and list expenses. This allows users to add new expenses with a description, amount, category, and time, as well as retrieve a list of all expenses.